requirements.txt
anyio==4.5.2
blinker==1.8.2
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.2
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
python-telegram-bot==21.6
sniffio==1.3.1
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.0.4
zipp==3.20.2
